Don't leak threads (and other things)
If we are closed from a separate thread, make sure to kill our main
thread.

If tubesock is closed from a different thread, then its own main thread
keeps sleeping on the select(2) forever. This leaks the thread and thus
memory. In ActiveRecord it also leaks the database connection owned by
the thread.